major change in IRCv3 logic
In past we used to strictly follow 3.1 standard which recommended to
process all CAPs before the actual registration to the network. This had
very bad side effect that we had to timeout / wait for server to respond
to CAP LS message, because non IRCv3 networks usually silently ignore
it. This made a big delay during registration to non IRCv3 networks,
causing timeouts on many of them and according to #ircv3 it's not really
needed as all IRCD implementations support IRCv3 negotiation after
login.

So now the logic is that we negotiate CAPs anytime, not only BEFORE the
registration, we simply process the registration ASAP and handle CAP
later.

It seems to work so far.